feat #49 : 인터뷰 영상, 트랙 정보 이동
인터뷰 영상과 트랙 정보의 위치를 Util 폴더로 변경하여 인터뷰 페이지 관련 코드를
정리하였습니다.
