+++ src/@Types/track.ts
@@ -1,0 +1,5 @@
+export interface TrackTypes {
+  id: string;
+  name: string;
+  link: string;
+}
