+++ src/@Types/video.ts
@@ -1,0 +1,8 @@
+export interface VideoTypes {
+  name: string;
+  job: string;
+  date: string;
+  thumbnail: string;
+  track: string;
+  url: string;
+}
