+++ src/Pages/Media/Interview/Components/Video/types.ts
@@ -1,12 +1,3 @@
 export interface VideoProps {
   id: string;
 }
-
-export interface VideoTypes {
-  name: string;
-  job: string;
-  date: string;
-  thumbnail: string;
-  track: string;
-  url: string;
-}
